Build out HRMS feature set on top of initial skeleton
Ship the end-to-end functional modules the frontend needs: attendance
(punch, regularization, geo-fencing, public holidays), employees
(extended profile, avatar upload, Excel bulk import), CTC-based
payroll, shift assignments, admin user/role management, dashboard,
and AI assistant wiring. Brings Flyway migrations V2-V11 to align
the schema with the entity model.

Infrastructure: add Maven wrapper (3.9.12) so contributors no longer
need a system Maven install, introduce S3/MinIO object storage for
avatars, fail-fast prod profile for missing secrets, and a device
API-key filter for fingerprint punch endpoints.

Also fixes a compile break: OfficeLocationServiceImpl,
PublicHolidayServiceImpl, and RegularizationServiceImpl were calling
a non-existent single-arg ResourceNotFoundException constructor.
Migrated them to the project-standard (resourceName, fieldName,
fieldValue) form used everywhere else.

+++ src/main/java/com/financebuddha/finbud/hrms/dto/shift/ShiftTypeResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
@@ -24,6 +25,6 @@ public class ShiftTypeResponse {
     private Integer gracePeriodMinutes;
     private List<Integer> weeklyOffDays;
     private Boolean isNightShift;
-    private Double overtimeThresholdHours;
+    private BigDecimal overtimeThresholdHours;
     private LocalDateTime createdAt;
 }
